+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2 - 2019-03-12
+
+* bug fix Android: Flutter页面没有全屏。模块开发时，返回键异常和不显示初始页面。
+* bug fix iOS: 非UINavigationController的rootViewController导致崩溃的bug。  
+
 ## 0.0.1 - 2019-03-11
 
 * 初步支持混合栈开发